program/requirements.txt
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
python-dotenv==1.2.1
